+++ example/lib/main.dart
@@ -1,14 +1,14 @@
 import 'dart:io';
 
 import 'package:flutter/material.dart';
-import 'package:than_pkg/enums/screen_orientation_types.dart';
 import 'package:than_pkg/than_pkg.dart';
-import 'package:than_pkg/types/src_dist_type.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
 
-  await ThanPkg.instance.init();
+  await ThanPkg.instance.init(
+    isShowDebugLog: true,
+  );
 
   runApp(const MyApp());
 }
@@ -36,74 +36,22 @@ class _MyAppState extends State<MyApp> {
       await ThanPkg.android.app.toggleKeepScreenOn(isKeep: true);
     }
 
-    MapServices.get<String>({}, ['name'], defaultValue: '');
-    MapServices.get<int>({}, ['num'], defaultValue: 1);
-    MapServices.get<double>({}, ['height'], defaultValue: 0.0);
-    //or
-    MapServices.get({}, ['keys'], defaultValue: ''); // -> type - auto cast
-    MapServices.get({
-      'main': {'key': 'than'}
-    }, [
-      'main',
-      'key'
-    ], defaultValue: 'default name');
+    // MapServices.get<String>({}, ['name'], defaultValue: '');
+    // MapServices.get<int>({}, ['num'], defaultValue: 1);
+    // MapServices.get<double>({}, ['height'], defaultValue: 0.0);
+    // //or
+    // MapServices.get({}, ['keys'], defaultValue: ''); // -> type - auto cast
+    // MapServices.get({
+    //   'main': {'key': 'than'}
+    // }, [
+    //   'main',
+    //   'key'
+    // ], defaultValue: 'default name');
 
     // ThanPkg.appUtil.getParseMinutes(minutes);
     // ThanPkg.appUtil.copyText(text);
     // ThanPkg.appUtil.pasteText();
     // await ThanPkg.appUtil.clearImageCache();
-  }
-
-  void _test() async {
-    try {
-      // extension
-      DateTime.now().toParseTime();
-      DateTime.now().toTimeAgo();
-      //double
-      0.0.toFileSizeLabel();
-      //FileSystemEntityExtension
-      // FileSystemEntityExtension.getName(withExt: false)
-
-      //PlatformExtension
-      PlatformExtension.isDesktop();
-      PlatformExtension.isMobile();
-
-      //StringExtension
-      "".toCaptalize();
-      "".getName();
-      "".getExt();
-      //TextEditingControllerExtension
-      TextEditingController().selectAll();
-
-      if (Platform.isAndroid) {
-        if (!await ThanPkg.android.permission.isStoragePermissionGranted()) {
-          await ThanPkg.android.permission.requestStoragePermission();
-          return;
-        }
-      }
-
-      final path = await ThanPkg.platform.getAppExternalPath();
-
-      await ThanPkg.platform.genPdfThumbnail(pathList: [
-        SrcDistType(src: '$path/Download/1-50.pdf', dist: '$path/test.png'),
-      ]);
-      await ThanPkg.platform.genVideoThumbnail(pathList: [
-        SrcDistType(
-            src: '$path/Download/catch.mp4', dist: '$path/catch-video.png'),
-      ]);
-
-      // final path = '/home/thancoder/Videos/【GMV】- BOSS B_TCH.mp4';
-      // await ThanPkg.platform.genVideoThumbnail(pathList: [
-      //   SrcDistType(src: path, dist: '/home/thancoder/Videos/test.png')
-      // ]);
-      // await ThanPkg.platform.genPdfThumbnail(pathList: [
-      //   SrcDistType(
-      //       src: '/home/thancoder/Downloads/mmbook/သရက်စိုက်ပျိုးနည်း.pdf',
-      //       dist: '/home/thancoder/Downloads/mmbook/test.png'),
-      // ]);
-    } catch (e) {
-      debugPrint(e.toString());
-    }
   }
 
   @override
@@ -121,10 +69,11 @@ class _MyAppState extends State<MyApp> {
         floatingActionButton: FloatingActionButton(
           // onPressed: _test,
           onPressed: () async {
+            // final res = await ThanPkg.platform.isInternetConnected();
             // final res = await ThanPkg.platform.getAppExternalPath();
-            final res = MapServices.getBool({'name': 134534}, ['name']);
-            ThanPkg.android.app
-                .requestOrientation(type: ScreenOrientationTypes.landscape);
+            // final res = MapServices.getBool({'name': 134534}, ['name']);
+            // ThanPkg.android.app
+            //     .requestOrientation(type: ScreenOrientationTypes.portrait);
             // final res = await ThanPkg.platform.getWifiAddressList();
             // final res = MapServices.getString({'name':5},['name'],defaultValue: 'default name');
             // final res = MapServices.getDouble({},['name'],defaultValue: 0.5);