+++ lib/types/src_dest_type.dart
@@ -1,0 +1,7 @@
+class SrcDestType {
+  String src;
+  String dest;
+  SrcDestType({required this.src, required this.dest});
+
+  Map<String, dynamic> toMap() => {'src': src, 'dist': dest};
+}